+++ MPC/config/rtps_relay_lib.mpb
@@ -1,5 +1,5 @@
 project: dcpslib, opendds_cxx11 {
   after += RtpsRelayLib
-  libs  += OpenDDS_RtpsRelay
-  includes += $(DDS_ROOT)/tools/rtpsrelay/lib
+  libs += OpenDDS_RtpsRelay
+  includes += $(DDS_ROOT)/tools
 }
