+++ tests/unit-tests/tools/rtpsrelay/lib/PartitionIndex.cpp
@@ -1,9 +1,10 @@
 #ifdef OPENDDS_HAS_CXX11
 
-#include <gtest/gtest.h>
-#include "tools/rtpsrelay/lib/PartitionIndex.h"
+#include <OpenDDS_RtpsRelayLib/PartitionIndex.h>
 
-#include "dds/DCPS/Service_Participant.h"
+#include <dds/DCPS/Service_Participant.h>
+
+#include <gtest/gtest.h>
 
 #include <iostream>
