+++ tools/OpenDDS_RtpsRelayLib/.gitignore
@@ -1,0 +1,8 @@
+/RelayC.h
+/RelayTypeSupport.idl
+/RelayTypeSupportC.cpp
+/RelayTypeSupportC.h
+/RelayTypeSupportC.inl
+/RelayTypeSupportImpl.cpp
+/RelayTypeSupportImpl.h
+/RelayTypeSupportS.h
