+++ tools/OpenDDS_RtpsRelayLib/RtpsRelayLib.mpc
@@ -1,4 +1,4 @@
-project(RtpsRelayLib): dcpslib, opendds_cxx11, install, msvc_bigobj {
+project: dcpslib, opendds_cxx11, install, msvc_bigobj {
   idlflags      += -Wb,export_macro=RtpsRelayLib_Export -Wb,export_include=export.h -SS
   dcps_ts_flags += -Wb,export_macro=RtpsRelayLib_Export -Wb,export_include=export.h
   dynamicflags  += RTPSRELAYLIB_BUILD_DLL
@@ -6,9 +6,5 @@ project(RtpsRelayLib): dcpslib, opendds_cxx11, install, msvc_bigobj {
 
   TypeSupport_Files {
     Relay.idl
-  }
-
-  specific {
-    install_headers =
   }
 }
