+++ tools/rtpsrelay/DomainStatisticsReporter.h
@@ -3,8 +3,8 @@
 
 #include "Config.h"
 
-#include "lib/RelayTypeSupportImpl.h"
-#include "lib/Utility.h"
+#include <OpenDDS_RtpsRelayLib/RelayTypeSupportImpl.h>
+#include <OpenDDS_RtpsRelayLib/Utility.h>
 
 #include <dds/DCPS/JsonValueWriter.h>
