+++ tools/rtpsrelay/RelayPartitionsListener.cpp
@@ -1,6 +1,6 @@
 #include "RelayPartitionsListener.h"
 
-#include "lib/RelayTypeSupportImpl.h"
+#include <OpenDDS_RtpsRelayLib/RelayTypeSupportImpl.h>
 
 #include <dds/DCPS/DCPS_Utils.h>
